requirements.txt
# Core framework
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-embeddings-mistralai>=0.1.0
llama-index-llms-mistralai>=0.1.0
llama-index-llms-ollama>=0.1.0

# Vector store
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Datasets
datasets==2.21.0  # Pinned for QASPER compatibility
huggingface-hub>=0.20.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
wikipedia-api>=0.6.0
httpx>=0.27.0
nltk>=3.8.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
